refactor: deduplicate initRepoWithRemoteAndBranch test helper into test-utils
